backend/requirements.txt
fastapi
uvicorn
python-multipart
demucs
pydub
ffmpeg-python
numpy
